Minesweeper field componentization
In this commit, the minesweeper field was turned into a component. Now, anyone using the component can define the width, height, and number of bombs.

Additionally, some errors were fixed, such as the use of fixed measurement units for style.

Finally, I would like to inform that SCSS and CSS files will soon be deprecated.