In this commit, I implemented several new features.
I fixed some bugs that I noticed in the code and added a filter to darken the background at the end of the game. Additionally, I resolved the issue with the game-over screen, which was not fully responsive.

Currently, I am presenting the game as a reusable component within the 'src' folder.

In the 'sample' folder, I provided an example of how to use the component and how to link a flag counter to it.

My next task (and possibly the last one) will be to address any remaining bugs in the game and then document the project.